# backend/swagger_documentation.yaml
openapi: 3.0.1
info:
  title: WaterFilterBusiness.API
  version: '1.0'
paths:
...
  /api/SmallInventoryRequests:
    post:
      tags:
        - SmallInventoryRequests
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryRequest_AddRequestModel'
          text/json:
            schema:
              $ref: '#/components/schemas/InventoryRequest_AddRequestModel'
          application/*+json:
            schema:
              $ref: '#/components/schemas/InventoryRequest_AddRequestModel'
      responses:
        '200':
          description: Success
    get:
      tags:
        - SmallInventoryRequests
      parameters:
        - name: page
          in: query
          style: form
          schema:
            type: integer
            format: int32
        - name: pageSize
          in: query
          style: form
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
...
components:
  schemas:
...
    InventoryRequest_AddRequestModel:
      required:
        - quantity
        - toolId
      type: object
      properties:
        toolId:
          type: integer
          format: int32
        quantity:
          type: integer
          format: int32
        requestNote:
          maxLength: 210
          minLength: 0
          type: string
          nullable: true
      additionalProperties: false
...